Moves camel_to_snake to .util
