+++ provenance_lib/tests/test_util.py
@@ -4,7 +4,41 @@ import zipfile
 
 from .test_parse import TEST_DATA
 from .testing_utilities import CustomAssertions
-from ..util import get_root_uuid, get_nonroot_uuid
+from ..util import get_root_uuid, get_nonroot_uuid, camel_to_snake
+
+
+class CamelToSnakeTests(unittest.TestCase):
+    def test_camel_to_snake(self):
+        some_types_n_formats = [
+            'Hierarchy',  # simple
+            'DistanceMatrix',  # compound
+            'Bowtie2Index',  # compound w numeral
+            'EMPPairedEndSequences',  # acronym
+            'BIOMV210DirFmt',  # acronym w numeral
+            'PCoAResults',  # weird acronym
+            'PairedEndFastqManifestPhred64V2',  # compound with acronym and num
+            'SampleData[Sequences]',  # bracket notation
+            'FeatureData[BLAST6]',  # bracket with acronym and numeral
+            'SampleData[DADA2Stats]',  # bracket complex acronym
+            'FeatureData[AlignedRNASequence]',  # bracket complex acronym
+            'List[FeatureTable[RelativeFrequency]]',  # made-up nested example
+        ]
+        exp = [
+            'hierarchy',
+            'distance_matrix',
+            'bowtie2_index',
+            'emp_paired_end_sequences',
+            'biomv210_dir_fmt',
+            'p_co_a_results',
+            'paired_end_fastq_manifest_phred64_v2',
+            'sample_data_sequences',
+            'feature_data_blast6',
+            'sample_data_dada2_stats',
+            'feature_data_aligned_rna_sequence',
+            'list_feature_table_relative_frequency',  # made-up nested example
+        ]
+        for og_str, exp_str in zip(some_types_n_formats, exp):
+            self.assertEqual(camel_to_snake(og_str), exp_str)
 
 
 class GetRootUUIDTests(unittest.TestCase):
