+++ provenance_lib/util.py
@@ -1,4 +1,5 @@
 import pathlib
+import re
 import zipfile
 
 # Alias string as UUID so we can specify types more clearly
@@ -31,3 +32,17 @@ def get_nonroot_uuid(fp: pathlib.Path) -> UUID:
     else:
         uuid = fp.parts[-2]
     return uuid
+
+
+def camel_to_snake(name: str) -> str:
+    """
+    There are more comprehensive and faster ways of doing this (incl compiling)
+    but it handles acronyms in semantic types nicely
+    e.g. EMPSingleEndSequences -> emp_single_end_sequences
+    c/o https://stackoverflow.com/a/1176023/9872253
+    """
+    # this will frequently be called on QIIME type expressions, so drop [ and ]
+    name = re.sub(r'[\[\]]', '', name)
+    # camel to snake
+    name = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
+    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', name).lower()
